+++ lib/src/common/widgets/circular_menu/menu.dart
@@ -87,8 +87,7 @@ class DraggableCircularMenu extends StatefulWidget {
   DraggableCircularMenuState createState() => DraggableCircularMenuState();
 }
 
-class DraggableCircularMenuState extends State<DraggableCircularMenu>
-    with SingleTickerProviderStateMixin {
+class DraggableCircularMenuState extends State<DraggableCircularMenu> with SingleTickerProviderStateMixin {
   late AnimationController _animationController;
 
   late int _itemsCount;
@@ -98,20 +97,16 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
   Timer? _fullHideTimer;
 
   final ValueNotifier<bool> _isDragging = ValueNotifier<bool>(false);
-  final ValueNotifier<Offset> _buttonPosition =
-      ValueNotifier<Offset>(const Offset(0, 100));
-  final ValueNotifier<double> _completeAngle =
-      ValueNotifier<double>(2 * math.pi);
+  final ValueNotifier<Offset> _buttonPosition = ValueNotifier<Offset>(const Offset(0, 100));
+  final ValueNotifier<double> _completeAngle = ValueNotifier<double>(2 * math.pi);
   final ValueNotifier<double> _initialAngle = ValueNotifier<double>(0);
   final ValueNotifier<bool> _startDragFromLeft = ValueNotifier<bool>(false);
-  final ValueNotifier<VisibleState> _visibleState =
-      ValueNotifier<VisibleState>(VisibleState.hidden);
+  final ValueNotifier<VisibleState> _visibleState = ValueNotifier<VisibleState>(VisibleState.hidden);
 
   @override
   void initState() {
     super.initState();
-    _animationController =
-        AnimationController(vsync: this, duration: widget.animationDuration);
+    _animationController = AnimationController(vsync: this, duration: widget.animationDuration);
     _animation = Tween(begin: 0.0, end: 1.0).animate(
       CurvedAnimation(
         parent: _animationController,
@@ -120,7 +115,7 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
       ),
     );
     _itemsCount = widget.items.length;
-    _startHideTimers();
+    _startHideTimers(context);
   }
 
   @override
@@ -145,7 +140,7 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
     return GestureDetector(
       onTap: () {
         if (_animationController.isCompleted) {
-          _closeMenu();
+          _closeMenu(context);
         }
       },
       child: Stack(
@@ -167,11 +162,10 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                 developer.log('ISpect: onTapOutside');
                 if (_animationController.status == AnimationStatus.dismissed) {
                   _visibleState.value = VisibleState.hidden;
-                } else if (_animationController.status ==
-                    AnimationStatus.completed) {
-                  _closeMenu();
+                } else if (_animationController.status == AnimationStatus.completed) {
+                  _closeMenu(context);
                 }
-                _startHideTimers();
+                _startHideTimers(context);
               },
               child: Stack(
                 children: [
@@ -196,21 +190,16 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                         _configureAlignmentBasedAngles(alignment);
                       });
                       return Positioned(
-                        left: _isInLeftSide(screenSize)
-                            ? _buttonPosition.value.dx
-                            : null,
+                        left: _isInLeftSide(screenSize) ? _buttonPosition.value.dx : null,
                         right: !_isInLeftSide(screenSize)
-                            ? screenSize.width -
-                                _buttonPosition.value.dx -
-                                widget.toggleButtonSize * 2
+                            ? screenSize.width - _buttonPosition.value.dx - widget.toggleButtonSize * 2
                             : null,
                         top: _buttonPosition.value.dy,
                         child: GestureDetector(
                           onPanStart: (details) {
                             if (details.globalPosition.dx < 80) {
                               _startDragFromLeft.value = true;
-                            } else if (details.globalPosition.dx >
-                                screenSize.width - 80) {
+                            } else if (details.globalPosition.dx > screenSize.width - 80) {
                               _startDragFromLeft.value = false;
                             }
                           },
@@ -229,7 +218,7 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                             if (_visibleState.value == VisibleState.active ||
                                 _visibleState.value == VisibleState.inactive) {
                               _snapButton(screenSize);
-                              _startHideTimers();
+                              _startHideTimers(context);
                             }
                             setState(() {
                               developer.log('ISpect: button rebuilded');
@@ -240,6 +229,7 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                             child: GestureDetector(
                               onTap: () {
                                 _closeMenu(
+                                  context,
                                   shouldReset: false,
                                 );
                                 if (widget.toggleButtonOnPressed != null) {
@@ -249,38 +239,28 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                               child: AnimatedContainer(
                                 duration: const Duration(milliseconds: 200),
                                 margin: EdgeInsets.zero,
-                                width:
-                                    _visibleState.value == VisibleState.hidden
-                                        ? 30
-                                        : widget.toggleButtonSize * 1.5,
+                                width: _visibleState.value == VisibleState.hidden ? 30 : widget.toggleButtonSize * 1.5,
                                 height: 70,
                                 decoration: BoxDecoration(
-                                  color:
-                                      _visibleState.value == VisibleState.hidden
-                                          ? widget.toggleButtonColor
-                                              .withOpacity(0.3)
-                                          : widget.toggleButtonColor,
+                                  color: _visibleState.value == VisibleState.hidden
+                                      ? widget.toggleButtonColor.withOpacity(0.3)
+                                      : widget.toggleButtonColor,
                                   borderRadius: BorderRadius.only(
-                                    topRight: _isInLeftSide(screenSize) ||
-                                            _isDragging.value
+                                    topRight: _isInLeftSide(screenSize) || _isDragging.value
                                         ? const Radius.circular(16)
                                         : Radius.zero,
-                                    bottomRight: _isInLeftSide(screenSize) ||
-                                            _isDragging.value
+                                    bottomRight: _isInLeftSide(screenSize) || _isDragging.value
                                         ? const Radius.circular(16)
                                         : Radius.zero,
-                                    topLeft: _isInRightSide(screenSize) ||
-                                            _isDragging.value
+                                    topLeft: _isInRightSide(screenSize) || _isDragging.value
                                         ? const Radius.circular(16)
                                         : Radius.zero,
-                                    bottomLeft: _isInRightSide(screenSize) ||
-                                            _isDragging.value
+                                    bottomLeft: _isInRightSide(screenSize) || _isDragging.value
                                         ? const Radius.circular(16)
                                         : Radius.zero,
                                   ),
                                 ),
-                                child: _visibleState.value ==
-                                        VisibleState.hidden
+                                child: _visibleState.value == VisibleState.hidden
                                     ? _isDragging.value
                                         ? const SizedBox()
                                         : Center(
@@ -288,21 +268,16 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                                               willChange: true,
                                               size: const Size(20, 70),
                                               painter: LineWithCurvePainter(
-                                                isInRightSide:
-                                                    _isInRightSide(screenSize),
-                                                color: widget
-                                                        .toggleButtonIconColor ??
-                                                    Colors.white,
+                                                isInRightSide: _isInRightSide(screenSize),
+                                                color: widget.toggleButtonIconColor ?? Colors.white,
                                               ),
                                             ),
                                           )
                                     : Center(
                                         child: AnimatedIcon(
-                                          icon: widget
-                                              .toggleButtonAnimatedIconData,
+                                          icon: widget.toggleButtonAnimatedIconData,
                                           size: widget.toggleButtonSize,
-                                          color: widget.toggleButtonIconColor ??
-                                              Colors.white,
+                                          color: widget.toggleButtonIconColor ?? Colors.white,
                                           progress: _animation,
                                         ),
                                       ),
@@ -326,8 +301,7 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
     _buttonPosition.value += details.delta;
     _isDragging.value = true;
     _buttonPosition.value = Offset(
-      _buttonPosition.value.dx
-          .clamp(0.0, screenSize.width - widget.toggleButtonSize * 2),
+      _buttonPosition.value.dx.clamp(0.0, screenSize.width - widget.toggleButtonSize * 2),
       _buttonPosition.value.dy.clamp(50, screenSize.height - 100),
     );
   }
@@ -353,18 +327,14 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
           builder: (_, __) => Positioned(
             left: _isInLeftSide(screenSize) ? _buttonPosition.value.dx : null,
             right: !_isInLeftSide(screenSize)
-                ? screenSize.width -
-                    _buttonPosition.value.dx -
-                    widget.toggleButtonSize * 2
+                ? screenSize.width - _buttonPosition.value.dx - widget.toggleButtonSize * 2
                 : null,
             top: _buttonPosition.value.dy,
             child: Transform.translate(
               offset: Offset.fromDirection(
                 _completeAngle.value == (2 * math.pi)
-                    ? (_initialAngle.value +
-                        (_completeAngle.value / _itemsCount) * index)
-                    : (_initialAngle.value +
-                        (_completeAngle.value / (_itemsCount - 1)) * index),
+                    ? (_initialAngle.value + (_completeAngle.value / _itemsCount) * index)
+                    : (_initialAngle.value + (_completeAngle.value / (_itemsCount - 1)) * index),
                 _animation.value * widget.radius,
               ),
               child: Transform.scale(
@@ -373,7 +343,9 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
                   angle: _animation.value * (math.pi * 2),
                   child: CircularMenuItemWidget(
                     item: item,
-                    closeMenu: _closeMenu,
+                    closeMenu: () {
+                      _closeMenu(context);
+                    },
                   ),
                 ),
               ),
@@ -397,54 +369,51 @@ class DraggableCircularMenuState extends State<DraggableCircularMenu>
     _isDragging.value = false;
   }
 
-  double _halfScreenWidth(Size screenSize) =>
-      screenSize.width / 2 - widget.toggleButtonSize / 2;
+  double _halfScreenWidth(Size screenSize) => screenSize.width / 2 - widget.toggleButtonSize / 2;
 
-  bool _isInLeftSide(Size screenSize) =>
-      _buttonPosition.value.dx < _halfScreenWidth(screenSize);
+  bool _isInLeftSide(Size screenSize) => _buttonPosition.value.dx < _halfScreenWidth(screenSize);
 
-  bool _isInRightSide(Size screenSize) =>
-      _buttonPosition.value.dx > _halfScreenWidth(screenSize);
+  bool _isInRightSide(Size screenSize) => _buttonPosition.value.dx > _halfScreenWidth(screenSize);
 
-  void _closeMenu({
+  void _closeMenu(
+    BuildContext context, {
     bool shouldReset = true,
   }) {
     if (_animationController.status == AnimationStatus.dismissed) {
       _animationController.forward();
       _visibleState.value = VisibleState.inactive;
       if (shouldReset) {
-        _resetHideTimers();
+        _resetHideTimers(context);
       }
     } else {
       _animationController.reverse();
     }
   }
 
-  void _startHideTimers() {
+  void _startHideTimers(BuildContext context) {
     _hideTimer = Timer(const Duration(seconds: 3), () {
-      if (_animationController.status == AnimationStatus.completed &&
-          !_isDragging.value) {
-        _closeMenu();
+      if (_animationController.status == AnimationStatus.completed && !_isDragging.value) {
+        _closeMenu(context);
       }
     });
     _fullHideTimer = Timer(const Duration(seconds: 6), () {
-      if (_animationController.status == AnimationStatus.dismissed &&
-          !_isDragging.value) {
+      if (_animationController.status == AnimationStatus.dismissed && !_isDragging.value) {
         _visibleState.value = VisibleState.hidden;
 
         Future.delayed(const Duration(milliseconds: 300), () {
+          if (!context.mounted) return;
           _snapButton(MediaQuery.sizeOf(context));
         });
       }
     });
   }
 
-  void _resetHideTimers() {
+  void _resetHideTimers(BuildContext context) {
     _hideTimer?.cancel();
     _fullHideTimer?.cancel();
     _visibleState.value = VisibleState.inactive;
 
-    _startHideTimers();
+    _startHideTimers(context);
   }
 
   Alignment _getAlignmentFromOffset(Offset offset, Size containerSize) {